CRITICAL FIX: Integer overflow causing negative ops/sec
- Bug: totalRange = rangePerThread * numThreads caused overflow
- For 900M × 8 = 7.2B, exceeds Int.MAX_VALUE (2.1B)
- Overflow wrapped to negative number → negative ops/sec
- Fix: Use Long instead of Int for totalRange calculation
- Changed: rangePerThread.toLong() * numThreads
- Result: Correct positive ops/sec values

+++ app/src/main/java/com/ivarna/finalbenchmark2/cpuBenchmark/algorithms/MultiCoreBenchmarks.kt
@@ -64,7 +64,9 @@ object MultiCoreBenchmarks {
 
                 // FIXED WORK PER CORE: Each thread processes the full range
                 val rangePerThread = params.primeRange
-                val totalRange = rangePerThread * numThreads
+                // CRITICAL: Use Long to prevent integer overflow
+                // For 900M × 8 cores = 7.2B, which exceeds Int.MAX_VALUE (2.1B)
+                val totalRange = rangePerThread.toLong() * numThreads
 
                 val (totalPrimes, timeMs) =
                         BenchmarkHelpers.measureBenchmark {
